Create equilibrium_in_3D.m
A slender light vertical rod is fixed in the ground at A.
At the top of the rod a rope AB is attached along which
a tensile force T = 10 kN is applied.
a) Calculate and plot the amount (magnitude) of the moment at O, M0 = |M 0| as a function of xB , −15m ≤ xB ≤ 15m, for zB = 9m.
b) Calculate for which values of xB where M0 < 100 kNm.
c) For what value of xB the minimum M0 is achieved. Also enter the value of M0 in this mode.