Create  Equilibrium_2D_friction.m
Equilibrium In 2D, friction.
A homogeneous cylinder with mass m = 30kg and radius r = 200mm rests against a rough vertical wall and a rough inclined surface as shown in the figure. The static coefficient of friction in both contacts is μs. Calculate the applied torque M required
to overcome the friction so that the cylinder begins to rotate.